Update 2 3/24/23
I just tested the code with group turn off for iris.csv and it worked.
I had to made a few changes, but I think the code is functional, though
I want to polish it up a bit more. I think I will include this in my
final project, so I need to figure out how to integrate it.